server/pyproject.toml
[project]
name = "finance-intel-hub-server"
version = "0.1.0"
description = "FinanceIntel Hub backend server"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "chromadb>=0.4.0",
    "openai>=1.12.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "apscheduler>=3.10.0",
    "pyyaml>=6.0.1",
    "jinja2>=3.1.0",
    "python-docx>=1.1.0",
    "pymupdf>=1.23.0",
    "firecrawl-py>=0.0.16",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
